Renamed Aux directory to ReactAux
