+++ src/components/studentTeamComponent/joinTeamRequest/JoinTeamRequestListComponent.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import JoinTeamRequestHelper from './joinTeamRequestHelper/JoinTeamRequestHelper'
-import Aux from '../../../hoc/Aux/Aux'
+import Aux from '../../../hoc/ReactAux/Aux'
 import '../../../assets/stylesheets/goldberg.css'
 
 const joinTeamRequestListComponent = (props) => {
